Move logEntryExit to logging, Move BaseProvider and INeeds to a new providerbase file
