+++ apis/taskcluster.py
@@ -4,7 +4,8 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-from components.utilities import logEntryExit, BaseProvider, INeedsCommandProvider, INeedsLoggingProvider
+from components.logging import logEntryExit
+from components.providerbase import BaseProvider, INeedsCommandProvider, INeedsLoggingProvider
 
 
 class TaskclusterProvider(BaseProvider, INeedsCommandProvider, INeedsLoggingProvider):
