+++ components/utilities.py
@@ -4,83 +4,7 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-import inspect
-
-from components.logging import LogLevel
-
-def logEntryExit(func):
-    def func_wrapper(*args, **kwargs):
-        obj = args[0]
-        assert 'logger' in dir(obj), "If @logEntryExit is applied to a class method, it must inherit INeedsLoggingProvider"
-        obj.logger.log("================================================", level=LogLevel.Debug)
-        obj.logger.log("Beginning %s" % func.__qualname__, level=LogLevel.Info)
-        obj.logger.log(" Arguments: %s" % str(args), level=LogLevel.Info)
-        ret = func(*args, **kwargs)
-        obj.logger.log("Ending %s" % func.__qualname__, level=LogLevel.Info)
-        return ret
-    return func_wrapper
-
 
 class Struct:
     def __init__(self, **entries):
         self.__dict__.update(entries)
-
-
-class BaseProvider:
-    """
-    All Providers - even test providers - must inherit from BaseProvider.
-    We expect an update_config method to be present on all providers, as it
-    will be called.
-
-    _update_config will most commonly be defined on INeeds interfaces, but can
-    be present on a derived class also.
-    """
-
-    def update_config(self, config):
-        """
-        update_config iterates through all the classes in the inheritence chain
-        and looks for an _update_config method on that class. If it is present,
-        it is called.
-
-        A provider may implement its own _update_config to perform actions for
-        itself.  Those actions are typically:
-          1. Calling update_config on a member provider or a collection of member
-             providers it has. Those members should inherit from BaseProvider
-             and note that we should only call update_config on them, this allows
-             the update_config call to succeed even if they have no INeeds
-             super-class or _update_config method.
-          2. Using its _update_config as a final init() method where all its
-             utility providers are available. (This is why we do the below in
-             reverse order.)
-
-
-        One wrinkle here is that if the derived class (the class of the variable;
-        i.e. not any superclasses) inherits from any class with _update_config,
-        and doesn't define its own, then one of the _update_config functions on
-        the base classes will be called twice. This shouldn't matter.
-        """
-        classes = self.__class__.mro()
-        classes.reverse()
-        for c in classes:
-            methods = inspect.getmembers(c, predicate=inspect.isfunction)
-            for m in methods:
-                if m[0] == '_update_config':
-                    m[1](self, config)
-
-
-class INeedsCommandProvider:
-    """
-    An interface class for Providers that need access to the command runner
-    """
-
-    def _update_config(self, config):
-        self.run = config['CommandProvider'].run
-
-
-class INeedsLoggingProvider:
-    """
-    An interface class for Providers that need access to logging
-    """
-
-    def _update_config(self, config):
-        self.logger = config['LoggingProvider'].bind_category(self.__class__.mro()[0].__name__)
